move AdalError to separate file
